+++ gluecodium/src/test/resources/smoke/constants/output/swift/smoke/ConstantsInterface.swift
@@ -19,7 +19,7 @@ public class ConstantsInterface {
     deinit {
         smoke_ConstantsInterface_release_handle(c_instance)
     }
-    public enum StateEnum : UInt32, CaseIterable {
+    public enum StateEnum : UInt32, CaseIterable, Codable {
         case off
         case on
     }