+++ gluecodium/src/test/resources/smoke/defaults/output/swift/smoke/TypesWithDefaults.swift
@@ -2,7 +2,7 @@
 //
 
 import Foundation
-public enum SomeEnum : UInt32, CaseIterable {
+public enum SomeEnum : UInt32, CaseIterable, Codable {
     case fooValue
     case barValue
 }