+++ gluecodium/src/test/resources/smoke/nesting/output/swift/smoke/FreeEnum.swift
@@ -2,7 +2,7 @@
 //
 
 import Foundation
-public enum FreeEnum : UInt32, CaseIterable {
+public enum FreeEnum : UInt32, CaseIterable, Codable {
     case foo
     case bar
 }