+++ gluecodium/src/test/resources/smoke/visibility/output/swift/smoke/PublicClass.swift
@@ -33,7 +33,7 @@ public class PublicClass {
     deinit {
         smoke_PublicClass_release_handle(c_instance)
     }
-    internal enum InternalEnum : UInt32, CaseIterable {
+    internal enum InternalEnum : UInt32, CaseIterable, Codable {
         case foo
         case bar
     }